# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "GameCoverageResponse": {
        "additionalProperties": false,
        "properties": {
          "analysis_coverage": {
            "enum": [
              "none",
              "partial",
              "complete"
            ],
            "title": "Analysis Coverage",
            "type": "string"
          },
          "analyzed_position_count": {
            "title": "Analyzed Position Count",
            "type": "integer"
          },
          "distinct_position_count": {
            "title": "Distinct Position Count",
            "type": "integer"
          },
          "preferred_coverage": {
            "enum": [
              "none",
              "partial",
              "complete"
            ],
            "title": "Preferred Coverage",
            "type": "string"
          },
          "preferred_position_count": {
            "title": "Preferred Position Count",
            "type": "integer"
          }
        },
        "required": [
          "distinct_position_count",
          "analyzed_position_count",
          "preferred_position_count",
          "analysis_coverage",
          "preferred_coverage"
        ],
        "title": "GameCoverageResponse",
        "type": "object"
      },
      "GameDetailOccurrenceResponse": {
        "additionalProperties": false,
        "properties": {
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "move_uci": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move Uci"
          },
          "ply": {
            "title": "Ply",
            "type": "integer"
          }
        },
        "required": [
          "ply",
          "fen",
          "move_uci"
        ],
        "title": "GameDetailOccurrenceResponse",
        "type": "object"
      },
      "GameDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "ended_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At Utc"
          },
          "game_uuid": {
            "format": "uuid",
            "title": "Game Uuid",
            "type": "string"
          },
          "occurrences": {
            "items": {
              "$ref": "#/components/schemas/GameDetailOccurrenceResponse"
            },
            "title": "Occurrences",
            "type": "array"
          },
          "opponent_chesscom_uuid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Chesscom Uuid"
          },
          "opponent_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Rating"
          },
          "original_pgn": {
            "title": "Original Pgn",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "started_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At Utc"
          },
          "termination_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termination Reason"
          },
          "time_class": {
            "anyOf": [
              {
                "enum": [
                  "bullet",
                  "blitz",
                  "rapid",
                  "daily"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Class"
          },
          "time_control": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Control"
          },
          "trainer_chesscom_uuid": {
            "format": "uuid",
            "title": "Trainer Chesscom Uuid",
            "type": "string"
          },
          "trainer_color": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Trainer Color",
            "type": "string"
          },
          "trainer_outcome": {
            "anyOf": [
              {
                "enum": [
                  "win",
                  "loss",
                  "draw"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trainer Outcome"
          },
          "trainer_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trainer Rating"
          }
        },
        "required": [
          "game_uuid",
          "source_url",
          "original_pgn",
          "trainer_color",
          "trainer_chesscom_uuid",
          "opponent_chesscom_uuid",
          "trainer_rating",
          "opponent_rating",
          "started_at_utc",
          "ended_at_utc",
          "trainer_outcome",
          "termination_reason",
          "time_control",
          "time_class",
          "occurrences"
        ],
        "title": "GameDetailResponse",
        "type": "object"
      },
      "GameSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "coverage": {
            "$ref": "#/components/schemas/GameCoverageResponse"
          },
          "deepest_opening": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpeningSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "ended_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At Utc"
          },
          "game_uuid": {
            "format": "uuid",
            "title": "Game Uuid",
            "type": "string"
          },
          "length_plies": {
            "title": "Length Plies",
            "type": "integer"
          },
          "occurrence_count": {
            "title": "Occurrence Count",
            "type": "integer"
          },
          "opponent_chesscom_uuid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Chesscom Uuid"
          },
          "opponent_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Rating"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "started_at_utc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At Utc"
          },
          "termination_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termination Reason"
          },
          "time_class": {
            "anyOf": [
              {
                "enum": [
                  "bullet",
                  "blitz",
                  "rapid",
                  "daily"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Class"
          },
          "time_control": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Control"
          },
          "trainer_chesscom_uuid": {
            "format": "uuid",
            "title": "Trainer Chesscom Uuid",
            "type": "string"
          },
          "trainer_color": {
            "enum": [
              "white",
              "black"
            ],
            "title": "Trainer Color",
            "type": "string"
          },
          "trainer_outcome": {
            "anyOf": [
              {
                "enum": [
                  "win",
                  "loss",
                  "draw"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trainer Outcome"
          },
          "trainer_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trainer Rating"
          }
        },
        "required": [
          "game_uuid",
          "source_url",
          "trainer_color",
          "trainer_chesscom_uuid",
          "opponent_chesscom_uuid",
          "trainer_rating",
          "opponent_rating",
          "started_at_utc",
          "ended_at_utc",
          "trainer_outcome",
          "termination_reason",
          "time_control",
          "time_class",
          "occurrence_count",
          "length_plies",
          "deepest_opening",
          "coverage"
        ],
        "title": "GameSummaryResponse",
        "type": "object"
      },
      "GamesErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "invalid_filter",
              "game_not_found",
              "games_unavailable",
              "unexpected_failure"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "GamesErrorResponse",
        "type": "object"
      },
      "GamesResponse": {
        "additionalProperties": false,
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages",
          "has_next"
        ],
        "title": "GamesResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpeningSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "eco": {
            "title": "Eco",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ply": {
            "title": "Ply",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "eco",
          "name",
          "ply"
        ],
        "title": "OpeningSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Chess Move Trainer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/games": {
      "get": {
        "operationId": "getGames",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "started_at_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Started At From"
            }
          },
          {
            "in": "query",
            "name": "started_at_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Started At To"
            }
          },
          {
            "in": "query",
            "name": "ended_at_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ended At From"
            }
          },
          {
            "in": "query",
            "name": "ended_at_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ended At To"
            }
          },
          {
            "in": "query",
            "name": "trainer_color",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trainer Color"
            }
          },
          {
            "in": "query",
            "name": "trainer_outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trainer Outcome"
            }
          },
          {
            "in": "query",
            "name": "termination_reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Termination Reason"
            }
          },
          {
            "in": "query",
            "name": "trainer_rating_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trainer Rating Min"
            }
          },
          {
            "in": "query",
            "name": "trainer_rating_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trainer Rating Max"
            }
          },
          {
            "in": "query",
            "name": "opponent_rating_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opponent Rating Min"
            }
          },
          {
            "in": "query",
            "name": "opponent_rating_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opponent Rating Max"
            }
          },
          {
            "in": "query",
            "name": "opponent_chesscom_uuid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opponent Chesscom Uuid"
            }
          },
          {
            "in": "query",
            "name": "time_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Time Class"
            }
          },
          {
            "in": "query",
            "name": "time_control",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Time Control"
            }
          },
          {
            "in": "query",
            "name": "opening_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opening Key"
            }
          },
          {
            "in": "query",
            "name": "opening_match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opening Match"
            }
          },
          {
            "in": "query",
            "name": "contains_fen",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contains Fen"
            }
          },
          {
            "in": "query",
            "name": "move_fen",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Move Fen"
            }
          },
          {
            "in": "query",
            "name": "move_uci",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Move Uci"
            }
          },
          {
            "in": "query",
            "name": "min_length_plies",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Length Plies"
            }
          },
          {
            "in": "query",
            "name": "max_length_plies",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Length Plies"
            }
          },
          {
            "in": "query",
            "name": "analysis_coverage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Analysis Coverage"
            }
          },
          {
            "in": "query",
            "name": "preferred_coverage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Preferred Coverage"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "started_at_desc",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Games",
        "tags": [
          "games"
        ]
      }
    },
    "/api/games/{game_uuid}": {
      "get": {
        "operationId": "getGame",
        "parameters": [
          {
            "in": "path",
            "name": "game_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Game",
        "tags": [
          "games"
        ]
      }
    },
...
  }
}